requirements.txt
cryptography>=42.0.0
pyOpenSSL>=24.0.0
charm-crypto-framework>=0.62